pyproject.toml
[project]
name = "volt-cli"
version = "0.1.1"
description = "An extremely fast template and stack manager for Python projects."
readme = "README.md"
authors = [
    { name = "Vishwa Kumaran", email = "vishwaa33400@gmail.com" }
]
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Framework :: FastAPI",
    "Topic :: Software Development :: Code Generators",
]
requires-python = ">=3.13"
dependencies = [
    "black>=25.9.0",
    "pydantic>=2.0.0",
    "questionary>=2.1.1",
    "tomli>=2.3.0",
    "tomli-w>=1.2.0",
    "typer>=0.19.2",
    "alembic>=1.13.0",
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["uv_build>=0.8.12,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "volt"
namespace = true

[dependency-groups]
dev = [
    "fastapi>=0.121.0",
    "httpx>=0.28.1",
    "pyinstaller>=6.17.0",
    "pytest>=8.4.2",
]

[project.scripts]
volt = "volt.cli:main"
